Improve project structure Add mamespace isolation
